refactor: move channels_v2/pipeline.py to channels/pipeline.py
