+++ apps/channels/channels_v2/channel_base.py
@@ -5,7 +5,7 @@ from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, ClassVar
 
 from apps.channels.capabilities import ChannelCapabilities
-from apps.channels.channels_v2.pipeline import MessageProcessingContext, MessageProcessingPipeline
+from apps.channels.pipeline import MessageProcessingContext, MessageProcessingPipeline
 from apps.channels.stages.core import (
     AttachmentHydrationStage,
     BotInteractionStage,
