+++ apps/channels/channels_v2/connect_channel.py
@@ -5,9 +5,9 @@ import logging
 from apps.channels.callbacks import ChannelCallbacks
 from apps.channels.capabilities import ChannelCapabilities, PlatformConsentConfig
 from apps.channels.channels_v2.channel_base import ChannelBase
-from apps.channels.channels_v2.pipeline import MessageProcessingContext
 from apps.channels.clients.connect_client import CommCareConnectClient
 from apps.channels.const import MESSAGE_TYPES
+from apps.channels.pipeline import MessageProcessingContext
 from apps.channels.sender import ChannelSender
 from apps.chat.exceptions import ChannelException
