+++ apps/channels/channels_v2/facebook_channel.py
@@ -11,9 +11,9 @@ from apps.channels.sender import ChannelSender
 if TYPE_CHECKING:
     from io import BytesIO
 
-    from apps.channels.channels_v2.pipeline import MessageProcessingContext
     from apps.channels.datamodels import BaseMessage
     from apps.channels.models import ExperimentChannel
+    from apps.channels.pipeline import MessageProcessingContext
     from apps.experiments.models import Experiment, ExperimentSession
     from apps.service_providers.messaging_service import MessagingService
     from apps.service_providers.speech_service import SynthesizedAudio
