+++ apps/channels/channels_v2/sureadhere_channel.py
@@ -10,8 +10,8 @@ from apps.channels.models import ChannelPlatform
 from apps.channels.sender import ChannelSender
 
 if TYPE_CHECKING:
-    from apps.channels.channels_v2.pipeline import MessageProcessingContext
     from apps.channels.models import ExperimentChannel
+    from apps.channels.pipeline import MessageProcessingContext
     from apps.experiments.models import Experiment, ExperimentSession
     from apps.service_providers.messaging_service import MessagingService
