+++ apps/channels/channels_v2/telegram_channel.py
@@ -14,9 +14,9 @@ from apps.channels import audio
 from apps.channels.callbacks import ChannelCallbacks
 from apps.channels.capabilities import ChannelCapabilities, PlatformConsentConfig
 from apps.channels.channels_v2.channel_base import ChannelBase
-from apps.channels.channels_v2.pipeline import MessageProcessingContext
 from apps.channels.const import MESSAGE_TYPES
 from apps.channels.datamodels import TelegramMessage
+from apps.channels.pipeline import MessageProcessingContext
 from apps.channels.sender import ChannelSender
 from apps.channels.stages.terminal import DeliveryErrorHandler, MessageDeliveryFailure
 from apps.experiments.models import ParticipantData
