+++ apps/channels/channels_v2/web_channel.py
@@ -8,9 +8,9 @@ from apps.channels.callbacks import ChannelCallbacks
 from apps.channels.capabilities import ChannelCapabilities
 from apps.channels.channels_v2.api_channel import NoOpSender
 from apps.channels.channels_v2.channel_base import ChannelBase
-from apps.channels.channels_v2.pipeline import MessageProcessingPipeline
 from apps.channels.const import MESSAGE_TYPES
 from apps.channels.models import ExperimentChannel
+from apps.channels.pipeline import MessageProcessingPipeline
 from apps.channels.stages.core import (
     BotInteractionStage,
     ChatMessageCreationStage,
