+++ apps/channels/stages/terminal.py
@@ -7,8 +7,8 @@ from typing import TYPE_CHECKING
 from django.utils import timezone
 
 from apps.annotations.models import TagCategories
-from apps.channels.channels_v2.pipeline import MessageProcessingContext
 from apps.channels.const import MESSAGE_TYPES
+from apps.channels.pipeline import MessageProcessingContext
 from apps.channels.stages.base import ProcessingStage
 from apps.channels.stages.core import RESET_COMMAND
 from apps.chat.models import ChatMessage, ChatMessageType
