+++ apps/channels/tests/channels/conftest.py
@@ -6,8 +6,8 @@ import pytest
 from apps.channels.callbacks import ChannelCallbacks
 from apps.channels.capabilities import ChannelCapabilities
 from apps.channels.channels_v2.channel_base import ChannelBase
-from apps.channels.channels_v2.pipeline import MessageProcessingContext
 from apps.channels.const import MESSAGE_TYPES
+from apps.channels.pipeline import MessageProcessingContext
 from apps.channels.sender import ChannelSender
 from apps.channels.tests.message_examples.base_messages import text_message
